added clap

+++ src/main.rs
@@ -2,20 +2,37 @@ use std::{fs, io};
 use crate::hash::process;
 mod search;
 mod hash;
-use std::env;
+
 use std::fs::File;
 use std::io::{BufWriter, Write};
+use clap::Parser;
 use crate::search::list_files;
 
+#[derive(Parser)]
+#[clap(author, version, about, long_about = None)]
+struct Cli {
+    /// Paths to files
+    #[clap(short, long, value_name = "PATH_TO_FILE", value_delimiter = ' ', num_args = 1..)]
+    paths: Vec<String>,
+
+    /// delete files
+    #[arg(short = 'd', long, action = clap::ArgAction::SetTrue)]
+    delete: bool,
+
+    /// verbose files
+    #[arg(short = 'v', long, action = clap::ArgAction::SetTrue)]
+    verbose: bool,
+}
+
 fn main() -> io::Result<()> {
-    let args: Vec<String> = env::args().collect();
-    let paths = args.into_iter().skip(1).collect::<Vec<String>>();
+    let cli = Cli::parse();
+
     println!("Looking in paths:");
-    for p in &paths {
+    for p in &cli.paths {
         println!("{p}");
     }
-    let files = list_files(paths)?;
-    let duplicates = process(files)?;
+    let files = list_files(cli.paths)?;
+    let duplicates = process(files, cli.verbose)?;
 
     let deleted_files = File::create("deleted_files.txt")?;
     let mut log = BufWriter::new(deleted_files);
@@ -25,7 +42,9 @@ fn main() -> io::Result<()> {
 
         for f in v.iter().skip(1) {
             write!(log, "{}\n", f)?;
-            delete_file(f)?;
+            if cli.delete {
+                delete_file(f)?;
+            }
             sum += 1;
         }
     }
